+++ Controls/_display/Collection.ts
@@ -157,23 +157,7 @@ export interface ISwipeConfig {
  * @property {AddPosition} [addPosition] Позиция редактирования по месту.
  * @property {Types/entity:Record} [item=undefined] Запись, которая будет запущена на редактирование при первой отрисовке списка.
  */
-/*
- * @typedef {Object} IEditingConfig
- * @property {Boolean} [editOnClick=false] If true, click on list item starts editing in place.
- * @property {Boolean} [autoAdd=false] If true, after the end of editing of the last list item, new item adds automatically and its editing begins.
- * @property {Boolean} [sequentialEditing=true] If true, after the end of editing of any list item other than the last, editing of the next list item starts automatically.
- * @property {Boolean} [toolbarVisibility=false] Determines whether buttons 'Save' and 'Cancel' should be displayed.
- * @property {AddPosition} [addPosition] Editing in place position.
- * @property {Types/entity:Record} [item=undefined] If present, editing of this item will begin on first render.
- */
-export interface IEditingConfig {
-    addPosition?: 'top'|'bottom';
-    toolbarVisibility?: boolean;
-    editOnClick?: boolean;
-    autoAdd?: boolean;
-    sequentialEditing?: boolean;
-    item?: CollectionItem<any>;
-}
+
 
 interface IUserStrategy<S, T> {
     strategy: new() => IItemsStrategy<S, T>;
